Cleanup Run.jl and move macro analysis to its own file
