+++ src/Pluto.jl
@@ -71,6 +71,7 @@ include("./analysis/DependencyCache.jl")
 include("./analysis/MoreAnalysis.jl")
 
 include("./evaluation/WorkspaceManager.jl")
+include("./evaluation/MacroAnalysis.jl")
 include("./packages/Packages.jl")
 include("./packages/PkgUtils.jl")
 include("./evaluation/Run.jl")
